requirements.txt
aniso8601==9.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-RESTful==0.3.9
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyngrok==5.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.1
